inicial

+++ src/main/java/com/everis/data/sumativa2/controllers/CategoriaController.java
@@ -1,5 +1,7 @@
 package com.everis.data.sumativa2.controllers;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,35 +21,54 @@ public class CategoriaController {
 	CategoriaService categoriaService;
 	
 	@RequestMapping("")
-	public String inicio(Model model) {
-		
+	public String inicio(Model model, HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		model.addAttribute("listadoCategorias",categoriaService.findAll());
 		
 		return "categoria.jsp";
+		}else
+			return"login.jsp";
 	}
 	
 	@RequestMapping("/insertar")
 	public String agregarCategoria(@RequestParam("nombre")String nombre,
-			Model model) {
+			Model model,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		Categoria categoria =  new Categoria();
 		categoria.setNombre(nombre);
 		
 		categoria = categoriaService.save(categoria);
 		model.addAttribute("listadoCategorias",categoriaService.findAll());
 		return "categoria.jsp";
+		}else {
+			return "login.jsp";
+		}
 	}
 	
 	@RequestMapping("/eliminar/{id}")
 	public String eliminarCategoria(@PathVariable("id") Long id,
-			Model model) {
+			Model model,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		categoriaService.deleteById(id);
 		model.addAttribute("listadoCategorias",categoriaService.findAll());
 		return "categoria.jsp";
-	}
+		}	else {
+			return "login.jsp";
+		}
+	
+}
 	
 	@RequestMapping("/editar/{id}")
 	public String editarCategoria(@RequestParam("newcat") String newcat,
-			@PathVariable("id")Long id, Model model) {
+			@PathVariable("id")Long id, Model model,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		Categoria cat=categoriaService.findById(id);
 		cat.setNombre(newcat);
 		categoriaService.save(cat);
@@ -55,6 +76,10 @@ public class CategoriaController {
 
 		return "categoria.jsp";
 		
-	}
+		}else {
+			return "login.jsp";
+		}
 
+	}
+	
 }
