+++ src/main/java/com/everis/data/sumativa2/controllers/ProductoController.java
@@ -2,7 +2,10 @@ package com.everis.data.sumativa2.controllers;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -18,6 +21,8 @@ import com.everis.data.sumativa2.services.ProductoService;
 @Controller
 @RequestMapping("/producto")
 public class ProductoController {
+	
+	private static final int CANT_PRODUCTOS = 2;
 
 	@Autowired
 	ProductoService productoService;
@@ -27,33 +32,56 @@ public class ProductoController {
 	
 	@RequestMapping("")
 	public String inicio(Model model,
-			@ModelAttribute("producto")Producto prod ) {
+			@ModelAttribute("producto")Producto prod ,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		
-		model.addAttribute("listadoProductos",productoService.findAll());
+		Page<Producto> productos= productoService.productosPaginados(0, CANT_PRODUCTOS);
+		
+		model.addAttribute("listadoProductos",productos);
 		model.addAttribute("listadoCategorias",categoriaService.findAll());
 		return "producto.jsp";
+		}else {
+			return "login.jsp";
+		}
 	}
 	
 	@RequestMapping("/insertar")
-	public String update(@ModelAttribute("producto")Producto prod, Model model ) {
+	public String update(@ModelAttribute("producto")Producto prod, Model model,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		productoService.save(prod);
 		model.addAttribute("listadoProductos",productoService.findAll());
 		model.addAttribute("listadoCategorias",categoriaService.findAll());
 
 		return "redirect:";
+		}else {
+			return "login.jsp";
+		}
 	}
 	@RequestMapping("/eliminar/{id}")
 	public String eliminarProducto(@PathVariable("id") Long id,
 			@ModelAttribute("producto")Producto prod,
-			Model model) {
+			Model model,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		productoService.deleteById(id);
 		model.addAttribute("listadoProductos",productoService.findAll());
 		model.addAttribute("listadoCategorias",categoriaService.findAll());
 		return "producto.jsp";
+		}else {
+			return "index.jsp";
+		}
 	}
 	
 	@RequestMapping("/editar/{id}")
-	public String editar(@PathVariable("id") Long id, Model model) {
+	public String editar(@PathVariable("id") Long id, Model model,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		Producto prod= productoService.findById(id);
 		
 		List<Categoria> listaCategorias =  categoriaService.findAll();
@@ -61,25 +89,45 @@ public class ProductoController {
 		model.addAttribute("producto", prod);
 		model.addAttribute("listaCategorias", listaCategorias);
 		return "editproducto.jsp";
+		}else {
+			return "login.jsp";
+		}
 	}
 	@RequestMapping("/update")
-	public String update(@ModelAttribute("producto")Producto prod ) {
+	public String update(@ModelAttribute("producto")Producto prod ,
+			HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		productoService.save(prod);
 		return "redirect:/producto";
+		}else {
+			return "login.jsp";
+		}
 	}
 	
 	@RequestMapping("/busqueda")
-	public String buscarProductos(Model model) {
+	public String buscarProductos(Model model, HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		return "productobusqueda.jsp";
+		}else {
+			return  "login.jsp";
+		}
 	}
 	
 	@RequestMapping("/busqueda/buscar")
 	public String busquedaProductos(@RequestParam("busqueda") String busqueda,
-			Model model) {
+			Model model, HttpSession session) {
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if(registrado==1) {
 		List<Producto> productos = productoService.FilterbyCategoria(busqueda);
 		System.out.println(productos);
 		model.addAttribute("listaProductos",productos);
 		return "productobusqueda.jsp";
+		}else {
+			return "login.jsp";
+					}
+		}
 	}
 	
-}
+
